backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
distro==1.9.0
exceptiongroup==1.2.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
lxml_html_clean==0.1.1
MarkupSafe==2.1.5
newspaper3k==0.2.8
nltk==3.8.1
openai==1.37.0
pillow==10.4.0
playwright==1.45.1
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
pyppeteer==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tinysegmenter==0.3
tldextract==5.1.2
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==1.26.19
websockets==10.4
Werkzeug==3.0.3
zipp==3.19.2
